adminserver/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.9
python-dotenv==1.0.0
pymongo==4.6.0
dnspython==2.4.2